Completed the reading route.
Button are working on bookID route.

Moved some code here and there, reducing boiler plate code.

Implemented a custom hook to make data fetching easy.

Have worked on making user experience amazing.

+++ server/controller/chapHandler.js
@@ -19,15 +19,15 @@ async function ChapCreate(req, res){
         const fileName = `${id}_${date.getTime()}`
         const filePath = `${Path}/${fileName}`
 
-        // const TitleExist = await Book.findOne({ 
-        //     $or: [ 
-        //         { 'publish[0]': req.body.title },
-        //         { 'draft[0]': req.body.title }
-        //     ]
-        // })
+        const TitleExist = await Book.findOne({
+            $or: [
+                { "draft.0": { $elemMatch: { $eq: req.body.title } } },
+                { "publish.0": { $elemMatch: { $eq: req.body.title } } }
+            ]
+        })
 
-        // if(TitleExist)
-        //     throw new Error("Chapter with same title already exists");
+        if(TitleExist)
+            throw new Error("Chapter with same title already exists");
 
         fs.writeFile(filePath, req.body.content, err => {
             if(err){
@@ -42,7 +42,7 @@ async function ChapCreate(req, res){
             };
 
             Book.findOneAndUpdate({ _id: id }, upData)
-                .then(data => res.send("Chapter Added"))
+                .then(data => res.status(200).send({msg: "Chapter Added", chpId: fileName }))
                 .catch(err => {
                     fs.unlink(filePath, (err) => {
                         //logger
@@ -67,4 +67,53 @@ async function OpenChap(req, res){
     })
 }
 
+async function EditChap(req, res) {
+    
+    try {
+        const id = ObjectId(req.params.id.split('_')[0]);
+        const appendType = req.query.q;
+
+        if (appendType !== 'publish' && appendType !== 'draft')
+            throw new Error("Invalid Request");
+
+        const date = new Date();
+        const fileName = `${id}_${date.getTime()}`
+        const filePath = `${Path}/${fileName}`
+
+        const TitleExist = await Book.findOne({
+            $or: [
+                { "draft.0": { $elemMatch: { $eq: req.body.title } } },
+                { "publish.0": { $elemMatch: { $eq: req.body.title } } }
+            ]
+        })
+
+        if (TitleExist)
+            throw new Error("Chapter with same title already exists");
+
+        fs.writeFile(filePath, req.body.content, err => {
+            if (err) {
+                //console.log(err) //logger
+                res.status(403).send("Error happened while trying to write chapter. Please try again later")
+            }
+
+            const upData = {
+                $push: {
+                    [appendType]: [req.body.title, fileName]
+                }
+            };
+
+            Book.findOneAndUpdate({ _id: id }, upData)
+                .then(data => res.status(200).send({ msg: "Chapter Added", chpId: fileName }))
+                .catch(err => {
+                    fs.unlink(filePath, (err) => {
+                        //logger
+                    })
+                    res.status(403).send("Couldn't add your chapter in the Collection.")
+                })
+        })
+    } catch (err) {
+        //console.log(err); //logger
+        res.status(403).send(err.message)
+    }
+}
 module.exports = { ChapCreate, OpenChap }
